requirements.txt
Flask==3.1.0
Flask-Cors==5.0.1
psycopg2-binary==2.9.10
pytest==8.3.5
python-dotenv==1.0.1
Werkzeug==3.1.3
gunicorn==23.0.0
